refactor PsifosModel and fixed routes bugs
